Architecture refactor to dependency injection model
